Move all service contracts to Api namespace - compliance with Magento coding standards
