+++ Api/TpayAliasServiceInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tpay\Magento2\Service;
+namespace Tpay\Magento2\Api;
 
 use Exception;
