+++ Notification/Strategy/BlikAliasNotificationProcessor.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Tpay\Magento2\Notification\Strategy;
 
-use Tpay\Magento2\Service\TpayAliasServiceInterface;
+use Tpay\Magento2\Api\Notification\Strategy\NotificationProcessorInterface;
+use Tpay\Magento2\Api\TpayAliasServiceInterface;
 
 class BlikAliasNotificationProcessor implements NotificationProcessorInterface
 {
