+++ Notification/Strategy/DefaultNotificationProcessor.php
@@ -4,6 +4,7 @@ namespace Tpay\Magento2\Notification\Strategy;
 
 use Exception;
 use Magento\Sales\Model\Order;
+use Tpay\Magento2\Api\Notification\Strategy\NotificationProcessorInterface;
 use Tpay\Magento2\Api\TpayConfigInterface;
 use Tpay\Magento2\Api\TpayInterface;
 use Tpay\Magento2\Service\TpayService;
